Move isolated runner into original crate as second binary, dynamically determine path to it
